+++ src/test/java/aut/funcional/testcases/cruceros/RBO_0013CB.java
@@ -1,0 +1,4 @@
+package aut.funcional.testcases.cruceros;
+
+public class RBO_0013CB {
+}
