+++ src/test/java/aut/funcional/testcases/cruceros/RBO_0016CM.java
@@ -1,0 +1,5 @@
+package aut.funcional.testcases.cruceros;
+
+public class RBO_0016CM {
+
+}
