--- src/test/java/aut/funcional/testcases/crucerosYTren/RBO_0014CB.java
@@ -1,4 +1,0 @@
-package aut.funcional.testcases.crucerosYTren;
-
-public class RBO_0014CB {
-}
