--- src/test/java/aut/funcional/testcases/crucerosYTren/RBO_0016CM.java
@@ -1,5 +1,0 @@
-package aut.funcional.testcases.crucerosYTren;
-
-public class RBO_0016CM {
-
-}
